% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censoring_adminstrative.R
\name{recruitment_uniform}
\alias{recruitment_uniform}
\alias{admin_censoring_time}
\alias{admin_censoring_events}
\title{Add recruitment time to Dataset}
\usage{
recruitment_uniform(dat, recruitment_until, recruitment_from = 0)

admin_censoring_time(dat, followup)

admin_censoring_events(dat, events)
}
\arguments{
\item{dat}{a simulated dataset}

\item{recruitment_until}{time of end of recruitment}

\item{recruitment_from}{time of start of recruitment (defaults to 0)}

\item{followup}{followup time}

\item{events}{}
}
\value{
Returns the dataset with added recruitment times.

Returns the dataset with administrative censoring after \code{follwup}, adds the
attribute \code{followup} with the followup time to the dataset.

Returns the dataset with administrative censoring after \code{events} events, adds
the attribute \code{followup} with the followup time to the dataset.
}
\description{
Add recruitment time to Dataset

Apply Administrative Censoring After Fixed Time

Apply Administrative Censoring After Fixed Number of Events
}
\details{
The Dataset hast to include a column \code{rec_time} containing the recruitment
time as well as the columns with the event times \code{t} and a column with the
event indicator \code{evt}.

Times and event indicaotrs for patients recruited after followup are set to
\code{NA}.

The Dataset hast to include a column \code{rec_time} containing the recruitment
time as well as the columns with the event times \code{t} and a column with the
event indicator \code{evt}.

Times and event indicaotrs for patients recruited after followup are set to
\code{NA}.
}
\section{Functions}{
\itemize{
\item \code{recruitment_uniform}: add recruitment time

\item \code{admin_censoring_time}: apply administrative censoring after fixed time

\item \code{admin_censoring_events}: apply administrative censoring after fixed number of events
}}

