% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_gehan_wilcoxon.R
\name{analyse_gehan_wilcoxon}
\alias{analyse_gehan_wilcoxon}
\title{Create Analyse function for Gehan Wilcoxon test}
\usage{
analyse_gehan_wilcoxon()
}
\value{
an analyse function that can be used in runSimulation
}
\description{
Create Analyse function for Gehan Wilcoxon test
}
\examples{
condition <- merge(
   assumptions_delayed_effect(),
   design_fixed_followup(),
   by=NULL
 ) |>
 head(1)
dat <- generate_delayed_effect(condition)
analyse_gehan_wilcoxon()(condition, dat)
}
