% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_group_sequential.R
\name{analyse_group_sequential}
\alias{analyse_group_sequential}
\title{Create Analyse Functions for Group Sequential Design}
\usage{
analyse_group_sequential(followup, followup_type, alpha, analyse_functions)
}
\arguments{
\item{followup}{followup events or time}

\item{followup_type}{"events" or "time"}

\item{alpha}{nominal alpha at each stage}

\item{analyse_functions}{analyse function or list of analyse functions}
}
\value{
an analyse function that can be used in runSimulation
}
\description{
Create Analyse Functions for Group Sequential Design
}
\details{
\code{followup}, \code{followup_type} and \code{alpha} are evaluated for every simulated
dataset, i.e. the arguments to the Analyse function are available,
expressions like \code{followup=c(condition$interim, condition$max_followup)} are
valid arguments.

analyse_functions should take arguments condition, dataset and fixed_objects
and return a list conatining p-value, number of patients and number of event
in the columsn \code{p}, \code{N_pat} and \code{N_evt}.
}
\examples{
# create a function to analyse after interim_events and maximum followup time
# given in the condition row of the design data.frame with given
# nominal alpha
analyse_maxcombo_sequential <- analyse_group_sequential(
  followup = c(condition$interim_events, condition$followup),
  followup_type = c("event", "time"),
  alpha = c(0.025, 0.05),
  analyse_functions = analyse_maxcombo
)
}
