% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_logrank.R
\name{analyse_logrank}
\alias{analyse_logrank}
\title{Analyse Dataset with the Logrank Test}
\usage{
analyse_logrank(condition, dat, fixed_objects = NULL)
}
\arguments{
\item{condition}{condition of the simulation}

\item{dat}{generated datasets}

\item{fixed_objects}{other constants}
}
\value{
a dataframe with the p-value of the logrank test in the column p
}
\description{
Analyse Dataset with the Logrank Test
}
\examples{
condition <- desing_skeleton_delayed_effect() |>
  head(1)
dat <- generate_delayed_effect(condition)
analyse_logrank(condition, dat)
}
