% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_logrank_fh_weights.R
\name{analyse_logrank_fh_weights}
\alias{analyse_logrank_fh_weights}
\title{Analyse Dataset with the Fleming Harrington weighted Logrank Test}
\usage{
analyse_logrank_fh_weights(rho, gamma)
}
\arguments{
\item{rho}{rho for the rho-gamma family of weights}

\item{gamma}{gamma for the rho-gamma family of weights}
}
\value{
a function with the arguments condition, dat and fixed_objects that
returns a dataframe with the p-value of the weighted logrank test in the
column p. See ?SimDesign::Analyse for details on the arguments condition,
dat, fixed_arguments.
}
\description{
Analyse Dataset with the Fleming Harrington weighted Logrank Test
}
\examples{
condition <- desing_skeleton_delayed_effect()[1, ]
dat <- generate_delayed_effect(condition)
# create two functions with different weights
analyse_01 <- analyse_logrank_fh_weights(rho=0, gamma=1)
analyse_10 <- analyse_logrank_fh_weights(rho=1, gamma=0)
# run the tests created before
analyse_01(condition, dat)
analyse_10(condition, dat)
}
