% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_maxcombo.R
\name{analyse_maxcombo}
\alias{analyse_maxcombo}
\title{Analyse Dataset with the Maxcombo Test}
\usage{
analyse_maxcombo()
}
\value{
an analyse function that returns a data.frame with the combined
p-value of the max combo test in the column p
}
\description{
Analyse Dataset with the Maxcombo Test
}
\examples{
condition <- merge(
    assumptions_delayed_effect(),
    design_fixed_followup(),
    by=NULL
  ) |>
  head(1)
dat <- generate_delayed_effect(condition)
analyse_maxcombo()(condition, dat)
}
