% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_modestly_weighted.R
\name{analyse_modelstly_weighted}
\alias{analyse_modelstly_weighted}
\title{Create Analyse function for the modestly weighted logrank test}
\usage{
analyse_modelstly_weighted(t_star)
}
\arguments{
\item{t_star}{parameter t* of the modestly weighted logrank test}
}
\value{
an analyse function that can be used in runSimulation
}
\description{
Create Analyse function for the modestly weighted logrank test
}
\examples{
condition <- merge(
   assumptions_delayed_effect(),
   design_fixed_followup(),
   by=NULL
 ) |>
 head(1)
dat <- generate_delayed_effect(condition)
analyse_modelstly_weighted(20)(condition, dat)
}
