% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_moderately_weighted.R
\name{analyse_moderately_weighted}
\alias{analyse_moderately_weighted}
\title{Create Analyse function for the moderately weighted logrank test}
\usage{
analyse_moderately_weighted()
}
\value{
an analyse function that can be used in runSimulation
}
\description{
Create Analyse function for the moderately weighted logrank test
}
\examples{
condition <- merge(
   assumptions_delayed_effect(),
   design_fixed_followup(),
   by=NULL
 ) |>
 head(1)
dat <- generate_delayed_effect(condition)
analyse_moderately_weighted()(condition, dat)
}
