% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_piecewise_exponential.R
\name{analyse_piecewise_exponential}
\alias{analyse_piecewise_exponential}
\title{Create Analyse function for piecewise exponential model}
\usage{
analyse_piecewise_exponential(times)
}
\arguments{
\item{times}{interval boundaries for the piecewise exponential model}
}
\value{
an analyse function that can be used in runSimulation
}
\description{
Create Analyse function for piecewise exponential model
}
\examples{
condition <- merge(
   assumptions_delayed_effect(),
   design_fixed_followup(),
   by=NULL
 ) |>
 head(1)
dat <- generate_delayed_effect(condition)
analyse_piecewise_exponential(times=c(90, 360))(condition, dat)
}
