% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_rmst.R
\name{analyse_rmst}
\alias{analyse_rmst}
\title{Analyse Dataset with an ANCOVA Type test on the RMST}
\usage{
analyse_rmst()
}
\value{
an analysis function that returns a data.frame with the p-value of
the test in the column p
}
\description{
Analyse Dataset with an ANCOVA Type test on the RMST
}
\examples{
condition <- merge(
    assumptions_delayed_effect(),
    design_fixed_followup(),
    by=NULL
  ) |>
  head(1)
dat <- generate_delayed_effect(condition)
analyse_rmst()(condition, dat)
}
