% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_rmst_diff.R
\name{analyse_rmst_diff}
\alias{analyse_rmst_diff}
\title{Analyse the Dataset using the difference in RMST}
\usage{
analyse_rmst_diff(condition, dat, fixed_objects = NULL)
}
\arguments{
\item{condition}{condition of the simulation}

\item{dat}{generated datasets}

\item{fixed_objects}{other constants}
}
\value{
a data frame with the columns
\itemize{
\item \code{p} p value of the test, see Details
\item \code{rmst_diff} estimated differnce in RMST
\item \code{rmst_diff_lower} adjusted lower bound of the confidence interval for differnce in RMST
\item \code{rmst_diff_upper} adjusted upper bound of the confidence interval for differnce in RMST
\item \code{N_pat} number of patients
\item \code{N_evt} number of events
}
}
\description{
Analyse the Dataset using the difference in RMST
}
\details{
The implementation from the nph package is used, see the documentation there
for details.
}
\examples{
condition <- merge(
    assumptions_delayed_effect(),
    design_fixed_followup(),
    by=NULL
  ) |>
  head(1)
dat <- generate_delayed_effect(condition)
analyse_rmst_diff(condition, dat)
}
\seealso{
\link\link{nph:nphparams}
}
