% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pch_functions_Rcpp.R
\name{check_tint_lambda}
\alias{check_tint_lambda}
\title{check if Tint and lambda arguments are ok}
\usage{
check_tint_lambda(Tint, lambda)
}
\arguments{
\item{Tint}{Tint argument to fast_... functions}

\item{lambda}{lambda argument to fast_... functions}
}
\value{
TRUE invisible
}
\description{
check if Tint and lambda arguments are ok
}
\details{
throws an error and of the following occur:
\itemize{
\item Tint is not sorted
\item any lambda are smaller than 0
\item the lengths of the vectors differ
}
}
\examples{
check_tint_lambda(c(0,1,2), c(0.1, 0.2, 0.1))
}
