% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination_tests_delayed.R
\docType{data}
\name{combination_tests_delayed}
\alias{combination_tests_delayed}
\title{Results of an example simulation}
\format{
a \code{tibble} as returned by \code{SimDesign::runSimulation}.
}
\usage{
combination_tests_delayed
}
\description{
Results of an example simulation study comparing the power of logrank
max-combo and modelstly weighted logrank test in differnt scenarios with
delayed onset of treatment effect.
}
\keyword{datasets}
