% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design_fixed_followup}
\alias{design_fixed_followup}
\title{Create a data.frame with an example fixed design}
\usage{
design_fixed_followup()
}
\value{
For design_fixed_followup: a design tibble with default values invisibly
}
\description{
Create a data.frame with an example fixed design
}
\details{
design_fixed_followup prints the code to generate a default
design tibble for use with generate_delayed_effect or other generate_...
functions and returns the evaluated code invisibly. This function is
intended to be used to copy paste the code and edit the parameters.
}
\section{Functions}{
\itemize{
\item \code{design_fixed_followup}: generate default fixed design
}}

\examples{
Design <- design_fixed_followup()
Design
}
