% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design_group_sequential}
\alias{design_group_sequential}
\title{Create a data.frame with an example group sequential design}
\usage{
design_group_sequential(print = interactive())
}
\arguments{
\item{print}{print code to generate parameter set?}
}
\value{
For design_group_sequential: a design tibble with default values invisibly
}
\description{
Create a data.frame with an example group sequential design
}
\details{
design_group_sequential generates a default design \code{data.frame} for
use with generate_delayed_effect or other generate_... functions. If print
is \code{TRUE} code to produce the template is also printed for copying, pasting
and editing by the user. (This is the default when run in an interactive
session.)
}
\section{Functions}{
\itemize{
\item \code{design_group_sequential()}: generate default group sequential design

}}
\examples{
Design <- design_group_sequential()
Design
}
