% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_crossing_hazards.R
\name{generate_crossing_hazards}
\alias{generate_crossing_hazards}
\alias{assumptions_crossing_hazards}
\alias{hr_after_crossing_from_gAHR}
\alias{true_summary_statistics_crossing_hazards}
\title{Generate Dataset with crossing hazards}
\usage{
generate_crossing_hazards(condition, fixed_objects = NULL)

assumptions_crossing_hazards()

hr_after_crossing_from_gAHR(design, target_gAHR)

true_summary_statistics_crossing_hazards(
  Design,
  cutoff_stats = 10,
  fixed_objects = NULL
)
}
\arguments{
\item{condition}{condition row of Design dataset}

\item{fixed_objects}{fixed objects of Design dataset}

\item{design}{design data.frame}

\item{target_gAHR}{target geometric average hazard ratio}

\item{Design}{Design data.frame for crossing hazards}

\item{cutoff_stats}{Cutoff time for rmst and average hazard ratios}

\item{fixed_objects=NULL}{fixed objects not used for now}
}
\value{
For generate_crossing_hazards: A dataset with the columns t (time) and trt
(1=treatment, 0=control), evt (event, currently TRUE for all observations)

For assumptions_crossing_hazards: a design tibble with default values invisibly

For hr_after_onset_from_gAHR: the design data.frame passed as
argument with the additional column hazard_trt.

For true_summary_statistics_crossing_hazards: the design data.frame
passed as argument with the additional columns:
\itemize{
\item \code{rmst_trt} rmst in the treatment group
\item \code{median_surv_trt} median survival in the treatment group
\item \code{rmst_ctrl} rmst in the control group
\item \code{median_surv_ctrl} median survial in the control group
\item \code{gAHR} geometric average hazard ratio
\item \code{AHR} average hazard ratio
}
}
\description{
Generate Dataset with crossing hazards

Create an empty assumtions data.frame for generate_crossing_hazards

Calculate hr after onset of treatment effect from gAHR

Calculate true summary statistics for scenarios with crossing hazards
}
\details{
Condidtion has to contain the following columns:
\itemize{
\item n_trt number of paitents in treatment arm
\item n_ctrl number of patients in control arm
\item crossing time of crossing of the hazards
\item hazard_ctrl hazard in the control arm = hazard before onset of treatment
effect
\item hazard_trt_before hazard in the treatment arm before onset of treatment effect
\item hazard_trt_after hazard in the treatment arm afert onset of treatment effect
}

If fixed_objects is given and contains an element \code{t_max}, then this is used
as the cutoff for the simulation used internally. If t_max is not given in
this way the 1-(1/10000) quantile of the survival distribution in the control
or treatment arm is used (which ever is larger).

assumptions_crossing_hazards prints the code to generate a default
design tibble for use with generate_crossing_hazards and returns the
evaluated code invisibly. This function is intended to be used to copy
paste the code and edit the parameters.
}
\section{Functions}{
\itemize{
\item \code{generate_crossing_hazards}: simulates a dataset with crossing
hazards

\item \code{assumptions_crossing_hazards}: generate default design tibble

\item \code{hr_after_crossing_from_gAHR}: Calculate hr after onset of treatment effect from gAHR

\item \code{true_summary_statistics_crossing_hazards}: calculate true summary statistics for crossing hazards
}}

\examples{
one_simulation <- merge(
    assumptions_crossing_hazards(),
    design_fixed_followup(),
    by=NULL
  ) |>
  head(1) |>
  generate_crossing_hazards()
head(one_simulation)
tail(one_simulation)
Design <- assumptions_crossing_hazards()
Design
my_design <- merge(
    assumptions_crossing_hazards(),
    design_fixed_followup(),
    by=NULL
  )
my_design$hr_trt <- NA
my_design <- hr_after_onset_from_gAHR(my_design, 0.8)
my_design
my_design <- merge(
    assumptions_crossing_hazards(),
    design_fixed_followup(),
    by=NULL
  )
my_design$follwup <- 15
my_design <- true_summary_statistics_crossing_hazards(my_design, cutoff_stats=my_design$followup)
my_design
}
