% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cure_mixture.R
\name{generate_cure_mixture}
\alias{generate_cure_mixture}
\alias{desing_skeleton_cure_mixture}
\title{Generate Dataset with mixture of cured patients}
\usage{
generate_cure_mixture(condition, fixed_objects = NULL)

desing_skeleton_cure_mixture()
}
\arguments{
\item{condition}{condition row of Design dataset}

\item{fixed_objects}{fixed objects of Design dataset}
}
\value{
For generate_cure_mixture: A dataset with the columns t (time) and trt
(1=treatment, 0=control), evt (event, currently TRUE for all observations)

For desing_skeleton_cure_mixture: a design tibble with default values invisibly
}
\description{
Generate Dataset with mixture of cured patients

Create design skeleton for use with generate_cure_mixture
}
\details{
Condidtion has to contain the following columns:
\itemize{
\item n_trt number of paitents in treatment arm
\item n_ctrl number of patients in control arm
\item hazard_ctrl hazard in the control arm
\item hazard_trt hazard in the treatment arm for not cured patients
\item hazard_cured hazard in for cured patients
\item cured_prop proportion of cured patients
\item t_max the maximal (cutoff) time
}

desing_skeleton_cure_mixture prints the code to generate a default
design tibble for use with generate_cure_mixture and returns the
evaluated code invisibly. This function is intended to be used to copy
paste the code and edit the parameters.
}
\section{Functions}{
\itemize{
\item \code{generate_cure_mixture}: simulates a dataset with a mixture of cured
patients

\item \code{desing_skeleton_cure_mixture}: generate default design tibble
}}

\examples{
one_simulation <- desing_skeleton_cure_mixture() |>
  head(1) |>
  generate_cure_mixture()
head(one_simulation)
tail(one_simulation)
Design <- desing_skeleton_cure_mixture()
Design
}
