% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_subgroup.R
\name{generate_subgroup}
\alias{generate_subgroup}
\alias{assumptions_subgroup}
\title{Generate Dataset with different treatment effect in subgroup}
\usage{
generate_subgroup(condition, fixed_objects = NULL)

assumptions_subgroup()
}
\arguments{
\item{condition}{condition row of Design dataset}

\item{fixed_objects}{fixed objects of Design dataset}
}
\value{
For generate_subgroup: A dataset with the columns t (time) and trt
(1=treatment, 0=control), evt (event, currently TRUE for all observations)

For assumptions_subgroup: a design tibble with default values invisibly
}
\description{
Generate Dataset with different treatment effect in subgroup

Create an empty assumtions data.frame for generate_cure_mixture
}
\details{
Condidtion has to contain the following columns:
\itemize{
\item n_trt number of paitents in treatment arm
\item n_ctrl number of patients in control arm
\item hazard_ctrl hazard in the control arm
\item hazard_trt hazard in the treatment arm for not cured patients
\item hazard_subgroup hazard in for cured patients
\item prevalence proportion of cured patients
}

assumptions_subgroup prints the code to generate a default
design tibble for use with generate_cure_mixture and returns the
evaluated code invisibly. This function is intended to be used to copy
paste the code and edit the parameters.
}
\section{Functions}{
\itemize{
\item \code{generate_subgroup}: simulates a dataset with a mixture of cured
patients

\item \code{assumptions_subgroup}: generate default assumptions tibble
}}

\examples{
one_simulation <- merge(
    assumptions_subgroup(),
    design_fixed_followup(),
    by=NULL
  ) |>
  head(1) |>
  generate_cure_mixture()
head(one_simulation)
tail(one_simulation)
Design <- assumptions_subgroup()
Design
}
