% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utilities.R
\name{generate_trial_inputs}
\alias{generate_trial_inputs}
\title{Generate a sidebar panel with select inputs for clinical trial parameters.}
\usage{
generate_trial_inputs(
  prefix,
  default_mts = 12,
  default_recruitment = 18,
  default_censoring = 0
)
}
\arguments{
\item{prefix}{A character string to use as a prefix for the input IDs.}

\item{default_mts}{The default value for the median survival time in the control arm.}

\item{default_recruitment}{The default value for the time to recruit target number of subjects.}

\item{default_censoring}{The default value for the proportion of subjects with random censoring.}
}
\value{
A sidebar panel with select inputs for clinical trial parameters.
}
\description{
Generate a sidebar panel with select inputs for clinical trial parameters.
}
\examples{
\dontrun{
# generate the sidebar panel with default values
sidebar_panel <- generate_trial_inputs("trial1_", default_mts = 12, default_recruitment = 18, default_censoring = 0)
}

}
