% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utilities.R
\name{interactive_filters}
\alias{interactive_filters}
\title{Interactive Filters}
\usage{
interactive_filters(prefix, input = NULL)
}
\arguments{
\item{prefix}{prefix to identify shiny input panel}

\item{input}{pass the shiny input object}
}
\description{
Utility function to define interactive (or not filters)
}
\details{
If input is set to NULL a character vector with static default filters will be returned,
this is useful for debugging without the need to rebuild the shiny document. Also input is not
available to the function unless it is defined within the markdown document (even if only sourced)
so you need to pass it from within the report.
}
