% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utilities.R
\name{plot_nph_curves}
\alias{plot_nph_curves}
\title{Plot Survival Curve}
\usage{
plot_nph_curves(data, type)
}
\arguments{
\item{data}{A data frame containing the parameters needed to generate the survival curve. The required columns depend on the type of survival curve to generate.}

\item{type}{A character string indicating the type of survival curve to generate. Valid values are 'delayed', 'progression', 'subgroup', and 'crossing'.}
}
\value{
A survival curve plot.
}
\description{
Plot a survival curve for a given set of parameters.
}
\examples{
\dontrun{
data <- data.frame(hazard_trt = 0.01, hazard_ctrl = 0.008, delay = 30, hazard_after_prog = 0.005, prog_rate_trt = 0.2, prog_rate_ctrl = 0.15, prevalence = 0.5)
plot_nph_curves(data = data, type = 'delay')
}
}
