% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utilities.R
\name{plot_parameter_scenario}
\alias{plot_parameter_scenario}
\title{Plot parameter scenarios}
\usage{
plot_parameter_scenario(
  data,
  parameters,
  parameter_x,
  filters = c("median_survival_ctrl == 12", "recruitment == 18", "censoring_prop == 0"),
  parameter_row = "effect_size_ph",
  parameter_col = "n_pat",
  parameters_steps = NULL,
  ...
)
}
\arguments{
\item{data}{tibble with simulation results in long format}

\item{parameters}{which parameter values to show on y-axis}

\item{parameter_x}{parameter to show on x-axis}

\item{filters}{character vector defining filters to apply to dataset (see Details)}

\item{parameter_row}{parameter to define row facets}

\item{parameter_col}{parameter to defin column facets}

\item{parameters_steps}{parameters across which x-scale is looped}

\item{...}{}
}
\description{
Nested loop plot of parameter scenarios
}
