% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utilities.R
\name{plot_sim_results}
\alias{plot_sim_results}
\title{Nested Loop Plot of Simulation Results}
\usage{
plot_sim_results(
  data,
  methods,
  parameter_x,
  parameter_y,
  filters = c("median_survival_ctrl == 12", "recruitment == 18", "censoring_prop == 0"),
  parameter_row = "effect_size_ph",
  parameter_col = "n_pat_design",
  parameters_steps = NULL,
  ...
)
}
\arguments{
\item{data}{tibble with simulation results in long format}

\item{methods}{methods for which operating characteristics should be plotted}

\item{parameter_x}{parameter to show on x-axis}

\item{parameter_y}{characteristic to show on y-axis}

\item{filters}{character vector defining filters to apply to dataset (see Details)}

\item{parameter_row}{parameter to define row facets}

\item{parameter_col}{parameter to defin column facets}

\item{parameters_steps}{parameters across which x-scale is looped}

\item{...}{additional options to looplot}
}
\description{
Generates a nested loop plot of simulation results
}
\details{
Parameters selected for x-axis, rows, cols, and steps need to be such that together with filters all remaining scenarios uniquely identify a point on the y-axis.
}
