% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{r2m}
\alias{r2m}
\alias{m2r}
\alias{m2d}
\alias{d2m}
\title{Utility Functions}
\usage{
r2m(lambda)

m2r(med)

m2d(mon)

d2m(day)
}
\arguments{
\item{lambda}{hazard rate}

\item{med}{median in months}

\item{mon}{time in months}

\item{day}{time in days}
}
\value{
median survival time in months (\code{r2m})

hazard rate per day (\code{m2r})

time in days (\code{m2d})

time in months (\code{d2m})
}
\description{
Some functions to convert between days and months and rates and medians.
}
\section{Functions}{
\itemize{
\item \code{r2m()}: daily rate to median in months

\item \code{m2r()}: median to months to daily rate

\item \code{m2d()}: months to days

\item \code{d2m()}: days to months

}}
\examples{
r2m(0.002)
m2r(12)
m2d(1)
d2m(31)
}
