% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_additional_sims.R
\name{rename_results_column}
\alias{rename_results_column}
\alias{rename_results_column_pattern}
\title{Rename Columns in Simulation Results and Update Attributes}
\usage{
rename_results_column(results, rename)

rename_results_column_pattern(results, pattern, replacement)
}
\arguments{
\item{results}{\code{SimDesign} object}

\item{rename}{named vector of new names}

\item{pattern}{regexp pattern as understood by \code{stringr::str_replace_all}}

\item{replacement}{replacement as understood by \code{stringr::str_replace_all}}
}
\value{
\code{SimDesign} object with updated column names
}
\description{
Rename Columns in Simulation Results and Update Attributes
}
\section{Functions}{
\itemize{
\item \code{rename_results_column()}: Rename Columns in Simulation Results

\item \code{rename_results_column_pattern()}: Rename Columns in Simulation Results by Pattern

}}
\examples{
\dontrun{
results2 <- rename_results_column(results, c("old_name"="new_name"))
}
\dontrun{
results2 <- rename_results_column_pattern(results, "old_name", "new_name")
}
}
