% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utilities.R
\name{tabulate_methods}
\alias{tabulate_methods}
\title{Tabulate Methods}
\usage{
tabulate_methods(data)
}
\arguments{
\item{data}{tibble with simulation results in long format}
}
\value{
tibble containing methods and what they provide
}
\description{
Return a table of all methods investigated in simulation results that
categorizes whether methods provide tests, estimators, and/or CIs
}
\examples{
\dontrun{
tabulate_methods(delayed_long)
}
}
