% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utilities.R
\name{tabulate_parameters}
\alias{tabulate_parameters}
\title{Tabulate Parameters}
\usage{
tabulate_parameters(data, parameters, .names = "Parameter")
}
\arguments{
\item{data}{tibble with simulation results (or parameter design) in long format}

\item{parameters}{which parameters to select for tabulation}

\item{.names}{column title for variables in table}
}
\value{
tibble containing selected parameters and corresponding unique values
}
\description{
Return a table with selected parameters and corresponding unique values
}
\examples{
\dontrun{
tabulate_parameters(sim_results_long, c("bias", "mse"))
}
}
