% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{wrap_all_in_trycatch}
\alias{wrap_all_in_trycatch}
\title{Wrap all functions in a list in tryCatch calls}
\usage{
wrap_all_in_trycatch(
  list_of_functions,
  error = function(e) {
     warning(e$message)
     NA
 }
)
}
\arguments{
\item{list_of_functions}{the list of functions to be wrapped}

\item{error}{the error function in the tryCatch call}
}
\value{
a list of functions
}
\description{
Wrap all functions in a list in tryCatch calls
}
\details{
SimDesign redraws data if one analysis function fails. This is not
only highly inefficient for large studies, but failure of a method is
informative and might be of interest. Moreover redrawing of data might
introduce bias if the failure of the method is not independent of the
parameter value, which would be a strong assumption.

To avoid redrawing data, we can catch all errors the analysis methods could
throw and return \code{NA} instead.

This is handled well by the summarise functions generated with
\code{create_summarise_function} other summarise functions might throw errors
when trying to \code{rbind} a data.frame to a scalar \code{NA} value. In this case
add another \code{error} argument. For example \verb{\\(e)\{NULL\}} could work in some
cases, in other cases you'll have to give a function that returns a
data.frame with the same columns as the analyse functions and only NA
values.
}
\examples{
Analyse <- list(
A = \(condition, dat, fixed_objects=NULL){
  if( runif(1) >= 0.5){
    counter_11(1)
    list(x=1, y=1)
  } else {
    counter_12(1)
    stop("asdf")
  }
},
B = \(condition, dat, fixed_objects=NULL){
  counter_13(1)
  list(z=1)
}
)

Analyse2 <- wrap_all_in_trycatch(Analyse)
Analyse2 <- wrap_all_in_trycatch(Analyse, error=\(e){NULL})
}
