% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_ahr.R
\name{analyse_ahr}
\alias{analyse_ahr}
\title{Analyse the dataset using extimators for the the average hazard ratio}
\usage{
analyse_ahr(max_time = NA, type = "AHR")
}
\arguments{
\item{max_time}{time for which the RMST is calculated}

\item{type=c("AHR", "gAHR")}{AHR for average hazard ratio gAHR for geometric average hazard ratio}
}
\value{
Returns an analysis function, that can be used in runSimulations
}
\description{
Analyse the dataset using extimators for the the average hazard ratio
}
\details{
The implementation from the nph package is used, see the documentation there
for details.

The data.frame returned by the created function includes the follwing
columns:
\itemize{
\item \code{p} p value of the test, see Details
\item \code{AHR}/\code{gAHR} estimated (geometric) average hazard ratio
\item \code{AHR_lower}/\code{gAHR_lower} unadjusted lower bound of the confidence interval for the (geometric) average hazard ratio
\item \code{AHR_upper}/\code{gAHR_upper} unadjusted upper bound of the confidence interval for the (geometric) average hazard ratio
\item \code{N_pat} number of patients
\item \code{N_evt} number of events
}
}
\examples{
condition <- merge(
    assumptions_delayed_effect(),
    design_fixed_followup(),
    by=NULL
  ) |>
  head(1)
dat <- generate_delayed_effect(condition)
analyse_ahr()(condition, dat)
analyse_ahr(type="gAHR")(condition, dat)
analyse_ahr(max_time=50, type="AHR")(condition, dat)
analyse_ahr(max_time=50, type="gAHR")(condition, dat)
}
\seealso{
\link{nph:nphparams}
}
