% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_coxph.R
\name{analyse_coxph}
\alias{analyse_coxph}
\title{Analyse Dataset with the Cox Protportional Hazards Model}
\usage{
analyse_coxph(condition, dat, fixed_objects = NULL)
}
\arguments{
\item{condition}{condition of the simulation}

\item{dat}{generated datasets}

\item{fixed_objects}{other constants}
}
\value{
a data frame with the columns
\itemize{
\item \code{p} p value of the score test
\item \code{coef} coefficient for \code{trt}
\item \code{hr} hazard ratio for \code{trt}
\item \code{confint} 95\% confidence intervall for the hazard ratio for \code{trt}
\item \code{N_pat} number of patients
\item \code{N_evt} number of events
}
}
\description{
Analyse Dataset with the Cox Protportional Hazards Model
}
\examples{
condition <- desing_skeleton_delayed_effect() |>
  head(1)
dat <- generate_delayed_effect(condition)
analyse_coxph(condition, dat)
}
