% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_coxph.R
\name{analyse_coxph}
\alias{analyse_coxph}
\title{Analyse Dataset with the Cox Protportional Hazards Model}
\usage{
analyse_coxph()
}
\value{
an analyse function that returns a list with the elements
\itemize{
\item \code{p} p value of the score test
\item \code{coef} coefficient for \code{trt}
\item \code{hr} hazard ratio for \code{trt}
\item \code{hr_lower} lower 95\% confidence intervall boundary for the hazard ratio for \code{trt}
\item \code{hr_upper}lower 95\% confidence intervall boundary for the hazard ratio for \code{trt}
\item \code{N_pat} number of patients
\item \code{N_evt} number of events
}
}
\description{
Analyse Dataset with the Cox Protportional Hazards Model
}
\examples{
condition <- merge(
    assumptions_delayed_effect(),
    design_fixed_followup(),
    by=NULL
  ) |>
  head(1)
dat <- generate_delayed_effect(condition)
analyse_coxph()(condition, dat)
}
