% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_logrank.R
\name{analyse_logrank}
\alias{analyse_logrank}
\title{Analyse Dataset with the Logrank Test}
\usage{
analyse_logrank()
}
\value{
an analysis function that returns a data.frame with the columns
\itemize{
\item \code{p} p-value of the logrank test
\item \code{N_pat} number of patients
\item \code{N_evt} number of events
}
}
\description{
Analyse Dataset with the Logrank Test
}
\examples{
condition <- merge(
    assumptions_delayed_effect(),
    design_fixed_followup(),
    by=NULL
  ) |>
  head(1)
dat <- generate_delayed_effect(condition)
analyse_logrank()(condition, dat)
}
