% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_milestone_survival.R
\name{analyse_milestone_survival}
\alias{analyse_milestone_survival}
\title{Analyse the Dataset using difference or quotient of milestone survival}
\usage{
analyse_milestone_survival(times, what = "quot", package = "nph")
}
\arguments{
\item{times}{followup times at which the the survival should be compared}

\item{what="quot"}{"quot" for quotient and "diff" for differnce of surival probabilities}

\item{package="nph"}{package to be used for computations: "nph" or "survival"}
}
\value{
Returns an analysis function, that can be used in runSimulations
}
\description{
Analyse the Dataset using difference or quotient of milestone survival
}
\details{
The implementation from the nph package is used, see the documentation there
for details.

The data.frame returned by the created function includes the follwing
columns:
\itemize{
\item \code{milestone_surv_ratio} / \code{milestone_surv_diff} ratio or differnce of survival probabilities
\item \code{times} followup times at which the the survival are compared
\item \code{N_pat} number of patients
\item \code{N_evt} number of events
}

If the package nph is used the following columns are also returned:
\itemize{
\item \code{p} p value for the H0 that the ratios are 1 or the differnce is 0 respectively
\item \code{milestone_surv_ratio_lower} / \code{milestone_surv_diff_lower} upper/lower CI for the estimate
\item \code{milestone_surv_ratio_upper} / \code{milestone_surv_diff_upper} upper/lower CI for the estimate
}
}
\examples{
condition <- merge(
    assumptions_delayed_effect(),
    design_fixed_followup(),
    by=NULL
  ) |>
  head(1)
dat <- generate_delayed_effect(condition)
analyse_milestone_survival(3:5)(condition, dat)
analyse_milestone_survival(3:5, what="diff")(condition, dat)
analyse_milestone_survival(3:5, what="quot", package="survival")(condition, dat)
analyse_milestone_survival(3:5, what="diff", package="survival")(condition, dat)
}
\seealso{
\link{nph:nphparams}
}
