% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_rmst.R
\name{analyse_rmst}
\alias{analyse_rmst}
\title{Analyse Dataset with an ANCOVA Type test on the RMST}
\usage{
analyse_rmst(condition, dat, fixed_objects = NULL)
}
\arguments{
\item{condition}{condition of the simulation}

\item{dat}{generated datasets}

\item{fixed_objects}{other constants}
}
\value{
a dataframe with the p-value of the test in the
column p
}
\description{
Analyse Dataset with an ANCOVA Type test on the RMST
}
\examples{
condition <- desing_skeleton_delayed_effect() |>
  head(1)
dat <- generate_delayed_effect(condition)
analyse_rmst(condition, dat)
}
