% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_rmst_diff.R
\name{analyse_rmst_diff}
\alias{analyse_rmst_diff}
\title{Analyse the Dataset using the difference in RMST}
\usage{
analyse_rmst_diff(max_time = NA)
}
\arguments{
\item{max_time}{time for which the RMST is calculated}
}
\value{
Returns an analysis function, that can be used in runSimulations
}
\description{
Analyse the Dataset using the difference in RMST
}
\details{
The implementation from the nph package is used, see the documentation there
for details.

The data.frame returned by the created function includes the follwing
columns:
\itemize{
\item \code{p} p value of the test, see Details
\item \code{rmst_diff} estimated differnce in RMST
\item \code{rmst_diff_lower} adjusted lower bound of the confidence interval for differnce in RMST
\item \code{rmst_diff_upper} adjusted upper bound of the confidence interval for differnce in RMST
\item \code{N_pat} number of patients
\item \code{N_evt} number of events
}
}
\examples{
condition <- merge(
    assumptions_delayed_effect(),
    design_fixed_followup(),
    by=NULL
  ) |>
  head(1)
dat <- generate_delayed_effect(condition)
analyse_rmst_diff()(condition, dat)
}
\seealso{
\link{nph:nphparams}
}
