% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_weibull.R
\name{analyse_weibull}
\alias{analyse_weibull}
\title{Analyse Dataset with Weibull Regression}
\usage{
analyse_weibull()
}
\value{
an analysis function that returns a data.frame with the columns
\itemize{
\item \code{est_med_trt} estimated median survival in the treatment arm
\item \code{est_med_ctrl} estimated median survival in the control arm
\item \code{N_pat} number of patients
\item \code{N_evt} number of events
}
}
\description{
Analyse Dataset with Weibull Regression
}
\examples{
condition <- merge(
    assumptions_delayed_effect(),
    design_fixed_followup(),
    by=NULL
  ) |>
  head(3) |>
  tail(1)
dat <- generate_delayed_effect(condition)
analyse_weibull()(condition, dat)
}
