% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design_fixed_followup}
\alias{design_fixed_followup}
\title{Create a data.frame with an example fixed design}
\usage{
design_fixed_followup(print = interactive())
}
\arguments{
\item{print}{print code to generate parameter set?}
}
\value{
For design_fixed_followup: a design tibble with default values invisibly
}
\description{
Create a data.frame with an example fixed design
}
\details{
design_fixed_followup generates a default design \code{data.frame} for
use with generate_delayed_effect or other generate_... functions. If print
is \code{TRUE} code to produce the template is also printed for copying, pasting
and editing by the user. (This is the default when run in an interactive
session.)
}
\section{Functions}{
\itemize{
\item \code{design_fixed_followup()}: generate default fixed design

}}
\examples{
Design <- design_fixed_followup()
Design
}
