% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design_group_sequential}
\alias{design_group_sequential}
\title{Create a data.frame with an example group sequential design}
\usage{
design_group_sequential()
}
\value{
For design_group_sequential: a design tibble with default values invisibly
}
\description{
Create a data.frame with an example group sequential design
}
\details{
design_group_sequential prints the code to generate a default
design tibble for use with generate_delayed_effect or other generate_...
functions and returns the evaluated code invisibly. This function is
intended to be used to copy paste the code and edit the parameters.
}
\section{Functions}{
\itemize{
\item \code{design_group_sequential}: generate default group sequential desing
}}

\examples{
Design <- design_group_sequential()
Design
}
