% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pch_functions_Rcpp.R
\name{fast_haz_fun}
\alias{fast_haz_fun}
\alias{fast_cumhaz_fun}
\alias{fast_cdf_fun}
\alias{fast_pdf_fun}
\alias{fast_surv_fun}
\alias{fast_quant_fun}
\title{Fast C implementation of hazard, cumulative hazard, ... for piecewise constant hazards}
\usage{
fast_haz_fun(Tint, lambda)

fast_cumhaz_fun(Tint, lambda)

fast_cdf_fun(Tint, lambda)

fast_pdf_fun(Tint, lambda)

fast_surv_fun(Tint, lambda)

fast_quant_fun(Tint, lambda)
}
\arguments{
\item{Tint}{time intervals on which the hazards are defined}

\item{lambda}{hazards in the time intervals}
}
\value{
A function with one parameter, a vector of times where the function should be evaluated.
}
\description{
Fast C implementation of hazard, cumulative hazard, ... for piecewise constant hazards
}
\section{Functions}{
\itemize{
\item \code{fast_haz_fun}: fast hazard function

\item \code{fast_cumhaz_fun}: fast cumulative hazard function

\item \code{fast_cdf_fun}: fast cumulative density function function

\item \code{fast_pdf_fun}: fast probability density function

\item \code{fast_surv_fun}: fast survival function

\item \code{fast_quant_fun}: fast quantile function
}}

\examples{
haz <- fast_haz_fun(c(0, 10, 20), c(10, 15, 7))
haz(seq(0, 30, by=0.1))
cumhaz <- fast_cumhaz_fun(c(0, 10, 20), c(10, 15, 7))
cumhaz(seq(0, 30, by=0.1))
cdf <- fast_cdf_fun(c(0, 10, 20), c(10, 15, 7))
cdf(seq(0, 30, by=0.1))
pdf <- fast_pdf_fun(c(0, 10, 20), c(10, 15, 7))
pdf(seq(0, 30, by=0.1))
surv <- fast_surv_fun(c(0, 10, 20), c(10, 15, 7))
surv(seq(0, 30, by=0.1))
quant <- fast_quant_fun(c(0, 10, 20), c(10, 15, 7))
quant(seq(0, 1, by=0.01))
}
