% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_delayed_effect.R
\name{generate_delayed_effect}
\alias{generate_delayed_effect}
\alias{desing_skeleton_delayed_effect}
\alias{prepare_design_delayed_effect_gAHR}
\title{Generate Dataset with delayed effect}
\usage{
generate_delayed_effect(condition, fixed_objects = NULL)

desing_skeleton_delayed_effect()

prepare_design_delayed_effect_gAHR(condition, target_gAHR)
}
\arguments{
\item{condition}{condition data.frame}

\item{fixed_objects}{fixed objects of Design dataset}

\item{target_gAHR}{target geometric average hazard ratio}
}
\value{
For generate_delayed_effect: A dataset with the columns t (time) and trt
(1=treatment, 0=control), evt (event, currently TRUE for all observations)

For desing_skeleton_delayed_effect: a design tibble with default values invisibly

For prepare_design_delayed_effect_gAHR: the condition
data.frame passed as argument with the additional column hazard_trt.
}
\description{
Generate Dataset with delayed effect

Create design skeleton for use with generate_delayed_effect

Calculate hr after onset of treatment effect from gAHR
}
\details{
Condidtion has to contain the following columns:
\itemize{
\item n_trt number of paitents in treatment arm
\item n_ctrl number of patients in control arm
\item delay time until onset of effect
\item hazard_ctrl hazard in the control arm = hazard before onset of treatment
effect
\item hazard_trt hazard in the treatment arm afert onset of treatment effect
}

desing_skeleton_delayed_effect prints the code to generate a default
design tibble for use with generate_delayed_effect and returns the
evaluated code invisibly. This function is intended to be used to copy
paste the code and edit the parameters.
}
\section{Functions}{
\itemize{
\item \code{generate_delayed_effect}: simulates a dataset with delayed
treatment effect

\item \code{desing_skeleton_delayed_effect}: generate default design tibble

\item \code{prepare_design_delayed_effect_gAHR}: Calculate hr after onset of treatment effect from gAHR
}}

\examples{
one_simulation <- desing_skeleton_delayed_effect() |>
  head(1) |>
  generate_delayed_effect()
head(one_simulation)
tail(one_simulation)
Design <- desing_skeleton_delayed_effect()
Design
my_condition <- desing_skeleton_delayed_effect()
my_condition$hr_trt <- NA
my_condition <- prepare_design_delayed_effect_gAHR(my_condition, 0.8)
my_condition
}
