% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_delayed_effect.R
\name{generate_delayed_effect}
\alias{generate_delayed_effect}
\alias{assumptions_delayed_effect}
\alias{hr_after_onset_from_gAHR}
\alias{hr_after_onset_from_PH_effect_size}
\alias{true_summary_statistics_delayed_effect}
\title{Generate Dataset with delayed effect}
\usage{
generate_delayed_effect(condition, fixed_objects = NULL)

assumptions_delayed_effect()

hr_after_onset_from_gAHR(design, target_gAHR, cutoff)

hr_after_onset_from_PH_effect_size(
  design,
  target_ph_power,
  followup,
  target_alpha = 0.05
)

true_summary_statistics_delayed_effect(
  Design,
  cutoff_stats = NA_real_,
  fixed_objects = NULL
)
}
\arguments{
\item{condition}{condition row of Design dataset}

\item{fixed_objects}{fixed objects of Design dataset}

\item{design}{design data.frame}

\item{target_gAHR}{target geometric average hazard ratio}

\item{cutoff}{time until which the gAHR should be calculated}

\item{target_ph_power}{target power under PH}

\item{followup}{followup time for the calculation of the effect size under PH}

\item{Design}{Design data.frame for delayed effect}

\item{target_alpha=0.05}{alpha level for the computation of the effect size under PH}

\item{cutoff_stats=NA_real_}{cutoff time, see details}

\item{fixed_objects=NULL}{additional settings, see details}
}
\value{
For generate_delayed_effect: A dataset with the columns t (time) and trt
(1=treatment, 0=control), evt (event, currently TRUE for all observations)

For assumptions_delayed_effect: a design tibble with default values invisibly

For hr_after_onset_from_gAHR: the design data.frame passed as
argument with the additional column hazard_trt.

For hr_after_onset_from_PH_effect_size: the design data.frame passed as
argument with the additional column hazard_trt.

For true_summary_statistics_delayed_effect: the design data.frame
passed as argument with the additional columns:
\itemize{
\item \code{rmst_trt} rmst in the treatment group
\item \code{median_surv_trt} median survival in the treatment group
\item \code{rmst_ctrl} rmst in the control group
\item \code{median_surv_ctrl} median survial in the control group
\item \code{gAHR} geometric average hazard ratio
\item \code{AHR} average hazard ratio
}
}
\description{
Generate Dataset with delayed effect

Create an empty assumtions data.frame for generate_delayed_effect

Calculate hr after onset of treatment effect from gAHR

Calculate hr after onset of treatment effect from target power under PH

Calculate true summary statistics for scenarios with delayed treatment effect
}
\details{
Condidtion has to contain the following columns:
\itemize{
\item n_trt number of paitents in treatment arm
\item n_ctrl number of patients in control arm
\item delay time until onset of effect
\item hazard_ctrl hazard in the control arm = hazard before onset of treatment
effect
\item hazard_trt hazard in the treatment arm afert onset of treatment effect
}

If fixed_objects is given and contains an element \code{t_max}, then this is used
as the cutoff for the simulation used internally. If t_max is not given in
this way the 1-(1/10000) quantile of the survival distribution in the control
or treatment arm is used (which ever is larger).

assumptions_delayed_effect prints the code to generate a default
design tibble for use with generate_delayed_effect and returns the
evaluated code invisibly. This function is intended to be used to copy
paste the code and edit the parameters.

This function works as follows: first the hazard ratio to archive
the target power under proportional hazards is calculated with Schönfeld's
formula. Next, the median survival for the treatment arm is calculated. And
finally the hazard rate for the treatment arm after onset of treatment
effect is calculated to yield the same median survival.

The if \code{fixed_objects} contains \code{t_max} then this value is used as the
maximum time to calculate function like survival, hazard, ... of the data
generating models. If this is not given \code{t_max} is choosen as the minimum of
the \code{1-(1/10000)} quantile of all survival distributions in the model.

\code{cutoff_stats} is the time used to calculate the statistics like average
hazard ratios and RMST, that are only calculated up to a certain point. It
defaults to \code{NA_real_} in which case the variable \code{followup} from the Design
dataset is used. If \code{followup} is also not set it uses \code{t_max}.
}
\section{Functions}{
\itemize{
\item \code{generate_delayed_effect}: simulates a dataset with delayed
treatment effect

\item \code{assumptions_delayed_effect}: generate default design tibble

\item \code{hr_after_onset_from_gAHR}: Calculate hr after onset of treatment effect from gAHR

\item \code{hr_after_onset_from_PH_effect_size}: Calculate hr after onset of treatment effect from gAHR

\item \code{true_summary_statistics_delayed_effect}: calculate true summary statistics for delayed effect
}}

\examples{
one_simulation <- merge(
    assumptions_delayed_effect(),
    design_fixed_followup(),
    by=NULL
  ) |>
  head(1) |>
  generate_delayed_effect()
head(one_simulation)
tail(one_simulation)
Design <- assumptions_delayed_effect()
Design
my_design <- merge(
    assumptions_delayed_effect(),
    design_fixed_followup(),
    by=NULL
  )
my_design$hazard_trt <- NA
my_design <- hr_after_onset_from_gAHR(my_design, 0.8, 200)
my_design
#' my_design <- merge(
    assumptions_delayed_effect(),
    design_fixed_followup(),
    by=NULL
  )
my_design$hazard_trt <- NA
my_design$hazard_ctrl <- 0.1
my_design <- hr_after_onset_from_PH_effect_size(my_design, 0.9, 200)
my_design
my_design <- merge(
    assumptions_delayed_effect(),
    design_fixed_followup(),
    by=NULL
  )
my_design$followup <- 15
my_design <- true_summary_statistics_delayed_effect(my_design)
my_design
}
