% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_delayed_effect.R
\name{generate_delayed_effect}
\alias{generate_delayed_effect}
\alias{desing_skeleton_delayed_effect}
\alias{hr_after_onset_from_gAHR}
\alias{true_summary_statistics_delayed_effect}
\title{Generate Dataset with delayed effect}
\usage{
generate_delayed_effect(condition, fixed_objects = NULL)

desing_skeleton_delayed_effect()

hr_after_onset_from_gAHR(design, target_gAHR)

true_summary_statistics_delayed_effect(
  Design,
  cutoff_stats = 10,
  fixed_objects = NULL
)
}
\arguments{
\item{condition}{condition row of Design dataset}

\item{fixed_objects}{fixed objects of Design dataset}

\item{design}{design data.frame}

\item{target_gAHR}{target geometric average hazard ratio}

\item{Design}{Design data.frame for delayed effect}

\item{cutoff_stats}{Cutoff time for rmst and average hazard ratios}

\item{fixed_objects=NULL}{fixed objects not used for now}
}
\value{
For generate_delayed_effect: A dataset with the columns t (time) and trt
(1=treatment, 0=control), evt (event, currently TRUE for all observations)

For desing_skeleton_delayed_effect: a design tibble with default values invisibly

For hr_after_onset_from_gAHR: the design data.frame passed as
argument with the additional column hazard_trt.

For true_summary_statistics_delayed_effect: the design data.frame
passed as argument with the additional columns:
\itemize{
\item \code{rmst_trt} rmst in the treatment group
\item \code{median_surv_trt} median survival in the treatment group
\item \code{rmst_ctrl} rmst in the control group
\item \code{median_surv_ctrl} median survial in the control group
\item \code{gAHR} geometric average hazard ratio
\item \code{AHR} average hazard ratio
}
}
\description{
Generate Dataset with delayed effect

Create design skeleton for use with generate_delayed_effect

Calculate hr after onset of treatment effect from gAHR

Calculate true summary statistics for scenarios with delayed treatment effect
}
\details{
Condidtion has to contain the following columns:
\itemize{
\item n_trt number of paitents in treatment arm
\item n_ctrl number of patients in control arm
\item delay time until onset of effect
\item hazard_ctrl hazard in the control arm = hazard before onset of treatment
effect
\item hazard_trt hazard in the treatment arm afert onset of treatment effect
}

If fixed_objects is given and contains an element \code{t_max}, then this is used
as the cutoff for the simulation used internally. If t_max is not given in
this way the 1-(1/10000) quantile of the survival distribution in the control
or treatment arm is used (which ever is larger).

desing_skeleton_delayed_effect prints the code to generate a default
design tibble for use with generate_delayed_effect and returns the
evaluated code invisibly. This function is intended to be used to copy
paste the code and edit the parameters.
}
\section{Functions}{
\itemize{
\item \code{generate_delayed_effect}: simulates a dataset with delayed
treatment effect

\item \code{desing_skeleton_delayed_effect}: generate default design tibble

\item \code{hr_after_onset_from_gAHR}: Calculate hr after onset of treatment effect from gAHR

\item \code{true_summary_statistics_delayed_effect}: calculate true summary statistics for delayed effect
}}

\examples{
one_simulation <- desing_skeleton_delayed_effect() |>
  head(1) |>
  generate_delayed_effect()
head(one_simulation)
tail(one_simulation)
Design <- desing_skeleton_delayed_effect()
Design
my_design <- desing_skeleton_delayed_effect()
my_design$hr_trt <- NA
my_design <- hr_after_onset_from_gAHR(my_design, 0.8)
my_design
my_design <- desing_skeleton_delayed_effect()
my_design <- true_summary_statistics_delayed_effect(my_design)
my_design
}
