% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_progression.R, R/generate_subgroup.R
\name{true_summary_statistics_progression}
\alias{true_summary_statistics_progression}
\alias{generate_subgroup}
\alias{assumptions_subgroup}
\alias{true_summary_statistics_subgroup}
\title{Calculate true summary statistics for scenarios with differential treatment effect in subgroup}
\usage{
true_summary_statistics_progression(
  Design,
  what = "os",
  cutoff_stats = 10,
  fixed_objects = NULL
)

generate_subgroup(condition, fixed_objects = NULL)

assumptions_subgroup()

true_summary_statistics_subgroup(
  Design,
  cutoff_stats = 10,
  fixed_objects = NULL
)
}
\arguments{
\item{Design}{Design data.frame for subgroup}

\item{cutoff_stats}{Cutoff time for rmst and average hazard ratios}

\item{fixed_objects}{fixed objects of Design dataset}

\item{condition}{condition row of Design dataset}

\item{what="os"}{True summary statistics for which estimand}

\item{fixed_objects=NULL}{fixed objects not used for now}
}
\value{
For true_summary_statistics_subgroup: the design data.frame
passed as argument with the additional columns:
\itemize{
\item \code{rmst_trt} rmst in the treatment group
\item \code{median_surv_trt} median survival in the treatment group
\item \code{rmst_ctrl} rmst in the control group
\item \code{median_surv_ctrl} median survial in the control group
\item \code{gAHR} geometric average hazard ratio
\item \code{AHR} average hazard ratio
}

For generate_subgroup: A dataset with the columns t (time) and trt
(1=treatment, 0=control), evt (event, currently TRUE for all observations)

For assumptions_subgroup: a design tibble with default values invisibly

For true_summary_statistics_subgroup: the design data.frame
passed as argument with the additional columns:
\itemize{
\item \code{rmst_trt} rmst in the treatment group
\item \code{median_surv_trt} median survival in the treatment group
\item \code{rmst_ctrl} rmst in the control group
\item \code{median_surv_ctrl} median survial in the control group
\item \code{gAHR} geometric average hazard ratio
\item \code{AHR} average hazard ratio
}
}
\description{
Calculate true summary statistics for scenarios with differential treatment effect in subgroup

Generate Dataset with different treatment effect in subgroup

Create an empty assumtions data.frame for generate_subgroup

Calculate true summary statistics for scenarios with differential treatment effect in subgroup
}
\details{
\code{what} can be \code{"os"} for overall survival and \code{"pfs"} for progression free
survival.

Condidtion has to contain the following columns:
\itemize{
\item n_trt number of paitents in treatment arm
\item n_ctrl number of patients in control arm
\item hazard_ctrl hazard in the control arm
\item hazard_trt hazard in the treatment arm for not cured patients
\item hazard_subgroup hazard in the subgroup in the treatment arm
\item prevalence proportion of cured patients
}

assumptions_subgroup prints the code to generate a default
design tibble for use with generate_subgroup and returns the
evaluated code invisibly. This function is intended to be used to copy
paste the code and edit the parameters.
}
\section{Functions}{
\itemize{
\item \code{true_summary_statistics_progression}: calculate true summary statistics for subgroup

\item \code{generate_subgroup}: simulates a dataset with a mixture of cured
patients

\item \code{assumptions_subgroup}: generate default assumptions tibble

\item \code{true_summary_statistics_subgroup}: calculate true summary statistics for subgroup
}}

\examples{
my_design <- merge(
    assumptions_progression(),
    design_fixed_followup(),
    by=NULL
  )
my_design$follwup <- 15
my_design_os  <- true_summary_statistics_subgroup(my_design, "os",  cutoff_stats=my_design$followup)
my_design_pfs <- true_summary_statistics_subgroup(my_design, "pfs", cutoff_stats=my_design$followup)
my_design_os
my_design_pfs
one_simulation <- merge(
    assumptions_subgroup(),
    design_fixed_followup(),
    by=NULL
  ) |>
  head(1) |>
  generate_subgroup()
head(one_simulation)
tail(one_simulation)
Design <- assumptions_subgroup()
Design
my_design <- merge(
    assumptions_subgroup(),
    design_fixed_followup(),
    by=NULL
  )
my_design <- true_summary_statistics_subgroup(my_design, cutoff_stats=my_design$followup)
my_design
}
