% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{r2m}
\alias{r2m}
\title{Compute median survival from hazard rate}
\usage{
r2m(lambda)
}
\arguments{
\item{lambda}{Hazard rate}
}
\value{
Median survival time in months
}
\description{
Compute the median survival time in months from the hazard rate \code{lambda}.
Assumes a constant hazard rate.
}
\examples{
r2m(0.01)

}
