% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_plots.R
\name{results_pivot_longer}
\alias{results_pivot_longer}
\alias{combined_plot}
\title{Functions for Plotting and Reporting Results}
\usage{
results_pivot_longer(data, exclude_from_methods = c("descriptive"))

combined_plot(
  data,
  methods,
  xvars,
  yvar,
  facet_x_vars = c(),
  facet_y_vars = c(),
  heights_plots = c(3, 1),
  scale_stairs = 0.75,
  grid_level = 2,
  scales = "fixed",
  hlines = numeric(0)
)
}
\arguments{
\item{data}{for combined_plto simulation results in long format, as returned by \code{results_pivot_longer}.}

\item{exclude_from_methods}{"methods" that should not be pivoted into long format}

\item{methods}{methods to include in the plot}

\item{xvars}{orderd vector of variable names to display on the x axis}

\item{yvar}{variable name of the variable to be displayed on the y axis (metric)}

\item{facet_x_vars}{vector of variable names to create columns of facets}

\item{facet_y_vars}{vector of variable names to create rows of facets}

\item{heights_plots}{relative heights of the main plot and the stairs on the bottom}

\item{scale_stairs}{height of the stairs for each variable between 0 and 1}

\item{grid_level}{depht of loops for which the grid-lines are drawn}

\item{scales}{passed on to facet_grid}

\item{hlines}{position of horizontal lines, passed as \code{yintercept} to
\code{geom_hline}}
}
\value{
dataset in long format with one row per method and scenario and one
column per metric

a ggplot/patchwork object conatining the plots
}
\description{
Functions for Plotting and Reporting Results
}
\details{
With \code{exclude_from_methods} descriptive statistics or results of
reference methods can be kept as own columns and used like the columns of
the simulation parameters.
}
\section{Functions}{
\itemize{
\item \code{results_pivot_longer}: pivot simulation results into long format

\item \code{combined_plot}: Nested Loop Plot with optional Facets
}}

\examples{
\dontrun{
  plot_data <- simulation_results |>
  results_pivot_longer
}
\dontrun{
  results_long <- results_wide |>
    results_pivot_longer()
  combined_plot(results_long, c("logrank", "max_combo"), c("effect_size_ph", "delay", "hazard_ctrl", "n_pat_design", "recruitment", "censoring_prop"), "rejection_0.05", grid_level=2)
}
}
