% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shhr_gg.R
\name{shhr_gg}
\alias{shhr_gg}
\title{Plot of survival, hazard and hazard ratio of two groups as a function of time using ggplot and patchwork}
\usage{
shhr_gg(
  A,
  B,
  main = NULL,
  sub = NULL,
  group_names = c("control", "treatment"),
  lab_time = "Days",
  lab_group = "Group",
  trafo_time = identity
)
}
\arguments{
\item{A}{mixpch object for group 1 (reference)}

\item{B}{mixpch object for group 2}

\item{main}{Title for the overall plot}

\item{sub}{Subtitle for the overall plot}

\item{group_names}{Group Names}

\item{lab_time}{Title for the time axis}

\item{lab_group}{Title group legend}

\item{trafo_time}{Function to transform time}
}
\value{
a \code{patchwork} object as defined in the patchwork package
}
\description{
Plot of survival, hazard and hazard ratio of two groups as a function of time using ggplot and patchwork
}
\examples{
library(ggplot2)
library(patchwork)
B <- pchaz(c(0, 10, 100), c(0.1, 0.05))
A <- pchaz(c(0, 100), c(0.1))
shhr_gg(A, B)
shhr_gg(A, B, lab_time="Months", trafo_time=d2m)
}
