% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNPH-package.R
\docType{package}
\name{SimNPH}
\alias{SimNPH}
\title{SimNPH: Simulate Non Proportional Hazards}
\description{
This package provides several functions to simulate survival data with non
proportional hazards using the general purpose simulation package SimDesign.
}
