% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_estimator}
\alias{summarise_estimator}
\title{Generic Summarise function for esitmators}
\usage{
summarise_estimator(est, real, lower = NULL, upper = NULL, name = NULL)
}
\arguments{
\item{est}{estimator, expression evaluated in results}

\item{real}{real summary statistic, expression evaluated in condition}

\item{lower}{lower CI, expression evaluated in results}

\item{upper}{upper CI, expression evaluated in results}

\item{name}{= NULL name for the summarise function, appended to the name of the analysis method in the final results}
}
\value{
A data frame with the columns
\itemize{
\item bias the bias with respect to the true value
\item var the variance
\item mse the mean squared error with respect to the true value
\item mae the mean absolute error with respect to the true value
\item coverage the coverage of the confidence interval
\item width the mean width of the confidence interval
}
}
\description{
Generic Summarise function for esitmators
}
\examples{
# generate the design matrix and append the true summary statistics
condition <- desing_skeleton_delayed_effect() |>
  tail(4) |>
  head(1) |>
  true_summary_statistics_delayed_effect(cutoff_stats = 15)

# create some summarise functions
summarise_all <- create_summarise_function(
  coxph=summarise_estimator(hr, gAHR, hr_lower, hr_upper, name="aAHR"),
  coxph=summarise_estimator(hr, hazard_trt/hazard_ctrl, hr_lower, hr_upper, name="HR"),
  coxph=summarise_estimator(exp(coef), gAHR),
  coxph=summarise_estimator(hr, NA_real_)
)

# runs simulations
sim_results <- runSimulation(
  design=condition,
  replications=10,
  generate=generate_delayed_effect,
  analyse=list(
    coxph=analyse_coxph
  ),
  summarise = summarise_all
)

# mse is missing for the summarise function in which the real value was NA
sim_results[, names(sim_results) |> grepl(pattern="mse")]
# but the variance can be estimated in all cases
sim_results[, names(sim_results) |> grepl(pattern="var")]
}
